26: project structure refactoring, create welcome-module, remove films-list module
