--- src/app/film-catalog/main/main.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'exp-main',
-  templateUrl: './main.component.html',
-  styleUrls: ['./main.component.scss']
-})
-export class MainComponent implements OnInit {
-
-  constructor() { }
-
-  public ngOnInit(): void { }
-
-}
