+++ src/app/films/components/details/details.component.scss
