+++ src/app/films/components/toolbar/toolbar.component.html
@@ -1,4 +1,3 @@
-
 <mat-toolbar class="menu-toolbar">
   <mat-toolbar-row>
