+++ src/app/shared/components/header/header.component.ts
@@ -8,7 +8,7 @@ import { Component, OnInit } from '@angular/core';
 export class HeaderComponent {
 
   links: object[] = [
-    { path: '/main', label: 'Main Page', active: 'button-active', icon: 'home'},
+    { path: '/welcome', label: 'Main Page', active: 'button-active', icon: 'home'},
     { path: '/films', label: 'All Films', active: 'button-active', icon: 'list_alt'},
     { path: '/actors', label: 'All Actors', active: 'button-active', icon: 'face'}
   ];
