+++ src/app/welcome/components/welcome/welcome.component.scss
