+++ src/app/welcome/components/welcome/welcome.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'exp-welcome',
+  templateUrl: './welcome.component.html',
+  styleUrls: ['./welcome.component.scss']
+})
+export class WelcomeComponent implements OnInit {
+
+  constructor() { }
+
+  public ngOnInit(): void { }
+
+}
