+++ src/app/welcome/welcome.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+
+
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatSelectModule } from '@angular/material/select';
+import { MatIconModule } from '@angular/material/icon';
+
+import { WelcomeComponent } from './components/welcome/welcome.component';
+
+
+@NgModule({
+  declarations: [
+    WelcomeComponent
+  ],
+  imports: [
+    CommonModule,
+    FormsModule,
+    MatTabsModule,
+    MatToolbarModule,
+    MatSelectModule,
+    MatIconModule
+  ]
+})
+export class WelcomeModule { }
